fix(dashboard0): refactor/reslove name collision in imports & rename submodules
