+++ src/dashboard/overlays/__init__.py
@@ -1,3 +1,3 @@
-from .settings import settings
+from ._settings import settings
 
 __all__ = ["settings"]
