+++ src/dashboard/tiles/__init__.py
@@ -1,4 +1,4 @@
-from .actions import actions
-from .messages import messages
-from .security import security
-from .account_info import account_info
+from ._actions import actions
+from ._messages import messages
+from ._security import security
+from ._account_info import account_info
